LPS-64316 Don't use registry-api in tests
https://issues.liferay.com/browse/LPS-64316
