+++ modules/apps/foundation/portal-search/portal-search/src/test/java/com/liferay/portal/search/internal/result/SearchResultUtilTest.java
@@ -15,11 +15,15 @@
 package com.liferay.portal.search.internal.result;
 
 import com.liferay.asset.kernel.AssetRendererFactoryRegistryUtil;
+import com.liferay.osgi.service.tracker.collections.map.ServiceReferenceMapper;
+import com.liferay.osgi.service.tracker.collections.map.ServiceTrackerMap;
+import com.liferay.osgi.service.tracker.collections.map.ServiceTrackerMapFactory;
 import com.liferay.portal.kernel.search.Document;
 import com.liferay.portal.kernel.search.DocumentImpl;
 import com.liferay.portal.kernel.search.Indexer;
 import com.liferay.portal.kernel.search.IndexerRegistry;
 import com.liferay.portal.kernel.search.SearchResult;
+import com.liferay.portal.kernel.search.SearchResultManager;
 import com.liferay.portal.kernel.search.Summary;
 import com.liferay.portal.kernel.search.SummaryFactory;
 import com.liferay.portal.kernel.search.result.SearchResultTranslator;
@@ -27,7 +31,6 @@ import com.liferay.portal.kernel.test.util.RandomTestUtil;
 import com.liferay.portal.kernel.util.StringPool;
 import com.liferay.portal.search.test.BaseSearchResultUtilTestCase;
 import com.liferay.portal.search.test.SearchTestUtil;
-import com.liferay.registry.collections.ServiceTrackerCollections;
 
 import java.util.List;
 import java.util.Locale;
@@ -36,12 +39,15 @@ import javax.portlet.PortletRequest;
 import javax.portlet.PortletResponse;
 
 import org.junit.Assert;
+import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
 import org.mockito.Matchers;
 import org.mockito.Mock;
 import org.mockito.Mockito;
+
+import org.osgi.framework.BundleContext;
 
 import org.powermock.core.classloader.annotations.PrepareForTest;
 import org.powermock.modules.junit4.PowerMockRunner;
@@ -51,11 +57,19 @@ import org.powermock.modules.junit4.PowerMockRunner;
  */
 @PrepareForTest(
 	{
-		AssetRendererFactoryRegistryUtil.class, ServiceTrackerCollections.class
+		AssetRendererFactoryRegistryUtil.class, ServiceTrackerMapFactory.class
 	}
 )
 @RunWith(PowerMockRunner.class)
 public class SearchResultUtilTest extends BaseSearchResultUtilTestCase {
+
+	@Before
+	@Override
+	public void setUp() throws Exception {
+		setUpServiceTrackerMap();
+
+		super.setUp();
+	}
 
 	@Test
 	public void testBlankDocument() {
@@ -177,6 +191,8 @@ public class SearchResultUtilTest extends BaseSearchResultUtilTestCase {
 
 		searchResultManagerImpl.setSummaryFactory(createSummaryFactory());
 
+		searchResultManagerImpl.activate(null);
+
 		return searchResultManagerImpl;
 	}
 
@@ -199,10 +215,32 @@ public class SearchResultUtilTest extends BaseSearchResultUtilTestCase {
 		return summaryFactoryImpl;
 	}
 
+	protected void setUpServiceTrackerMap() {
+		mockStatic(ServiceTrackerMapFactory.class, Mockito.CALLS_REAL_METHODS);
+
+		stub(
+			method(
+				ServiceTrackerMapFactory.class, "openSingleValueMap",
+				BundleContext.class, Class.class, String.class,
+				ServiceReferenceMapper.class)
+		).toReturn(
+			_serviceTrackerMap
+		);
+
+		when(
+			_serviceTrackerMap.getService(Mockito.anyString())
+		).thenReturn(
+			null
+		);
+	}
+
 	@Mock
 	private Indexer<Object> _indexer;
 
 	@Mock
 	private IndexerRegistry _indexerRegistry;
+
+	@Mock
+	private ServiceTrackerMap<String, SearchResultManager> _serviceTrackerMap;
 
 }
